Implemented pool for shared background queue dispatchers
Also using WeakReferences in order to allow the state machine and
dispatchers to be garbage collected as soon as there are no hard
references to them.

A known issue is that the state machine won't be garbage collected
directly when it is unreferenced, rather when all queued tasks
have finished. The issue is described in BackgroundQueueDispatcher.