Added support for auto connections that trigger on state change
If ! is specified as signal for a connection with a from and to
state it will be triggered at the end of a state transition. This
enables for automatic state transitions to occur. This is a nice
way to re-use code. If a state is reachable from multiple other
states you can point them to a pre-state and then connect the
pre-state with the actual state using an auto connection. Write
the shared code in the auto connection method and you're good
to go.