Added global setLogger feature and fixed warning for onEnter and onExit
The class AnnoStateMachine was introduced. It has a static setLogger
method that can be used to set the globally used logger.

The onEnter and onExit methods always trigger regardless of return
value from previous calls. The dispatch code however returned
null for the guard state instead of true which generated a warning.