Added platform addon support
The Android lib now sets an Android specific logger instance.

If a platform specific lib has addons it must create an Addons.java
file in the com.jayway.annostatemachine package and inherit AddonRepo.

If such a file exists it will be loaded and the addons that it exposes
will be initialized. This happens when the first state machine
implementation class is intantiated and calls Config.get().