Improved error handling in processing step
We always generate a class so that imports or method calls do not fail in
client code. We output warnings instead of errors so that we do not
get unnecessary compiler errors that clog up the log. If a state machine
is incomplete the annotation @IncompleteStateMachine is added and a separate
processor has been added for that annotation. The annotation is added by the
StateMachine annotation processor which means that the IncompleteStateMachine
annotation processor can safely fail the build.